+++ scanpipe/tests/pipes/test_purldb.py
@@ -21,6 +21,7 @@
 # Visit https://github.com/nexB/scancode.io for support and download.
 
 import io
+import json
 from pathlib import Path
 from unittest import mock
 
@@ -35,6 +36,7 @@ from scanpipe.models import Run
 from scanpipe.pipes import purldb
 from scanpipe.tests import dependency_data2
 from scanpipe.tests import dependency_data3
+from scanpipe.tests import make_package
 from scanpipe.tests import package_data1
 
 
@@ -234,3 +236,59 @@ class ScanPipePurlDBTest(TestCase):
         scannable_uri_uuid = "52b2930d-6e85-4b3e-ba3e-17dd9a618650"
         project_name = purldb.create_project_name(download_url, scannable_uri_uuid)
         self.assertEqual("httpsregistrynpmjsorgasdf-asdf-101tgz-52b2930d", project_name)
+
+    @mock.patch("scanpipe.pipes.purldb.get_package_by_purl")
+    def test_scanpipe_pipes_purldb_enrich_package(self, mock_get_package_by_purl):
+        package1 = make_package(self.project1, package_url="pkg:npm/csvtojson@2.0.10")
+
+        mock_get_package_by_purl.return_value = {}
+        updated_fields = purldb.enrich_package(package=package1)
+        self.assertIsNone(updated_fields)
+
+        purldb_entry_file = self.data / "purldb" / "csvtojson-2.0.10.json"
+        purldb_entry = json.loads(purldb_entry_file.read_text())
+        mock_get_package_by_purl.return_value = purldb_entry
+        updated_fields = purldb.enrich_package(package=package1)
+        self.assertTrue(updated_fields)
+        self.assertIn("homepage_url", updated_fields)
+
+        package1.refresh_from_db()
+        self.assertTrue(package1.extra_data.get("enrich_with_purldb"))
+        self.assertEqual(purldb_entry.get("sha1"), package1.sha1)
+        self.assertEqual(purldb_entry.get("sha256"), package1.sha256)
+        self.assertEqual(purldb_entry.get("copyright"), package1.copyright)
+
+    @mock.patch("scanpipe.pipes.purldb.get_package_by_purl")
+    def test_scanpipe_pipes_purldb_enrich_discovered_packages(
+        self, mock_get_package_by_purl
+    ):
+        package1 = make_package(self.project1, package_url="pkg:npm/csvtojson@2.0.10")
+
+        mock_get_package_by_purl.return_value = {}
+        buffer = io.StringIO()
+        updated_package_count = purldb.enrich_discovered_packages(
+            project=self.project1,
+            logger=buffer.write,
+        )
+        self.assertEqual(0, updated_package_count)
+        expected_log = buffer.getvalue()
+        self.assertIn("0 discovered packages enriched with the PurlDB.", expected_log)
+
+        purldb_entry_file = self.data / "purldb" / "csvtojson-2.0.10.json"
+        purldb_entry = json.loads(purldb_entry_file.read_text())
+        mock_get_package_by_purl.return_value = purldb_entry
+        buffer = io.StringIO()
+        updated_package_count = purldb.enrich_discovered_packages(
+            project=self.project1,
+            logger=buffer.write,
+        )
+        self.assertEqual(1, updated_package_count)
+        package1.refresh_from_db()
+        self.assertEqual(purldb_entry.get("sha1"), package1.sha1)
+        self.assertEqual(purldb_entry.get("sha256"), package1.sha256)
+        self.assertEqual(purldb_entry.get("copyright"), package1.copyright)
+        self.assertTrue(package1.extra_data.get("enrich_with_purldb"))
+
+        expected_log = buffer.getvalue()
+        self.assertIn("pkg:npm/csvtojson@2.0.10 ['release_date'", expected_log)
+        self.assertIn("1 discovered package enriched with the PurlDB.", expected_log)